allow to modify a row during diving if not in LP
Creating a row and adding it, e.g., to a cutpool, during diving
did not seem possible in debug mode due to asserts in lp.c.
This commit changes the asserts so that it is allowed to modify a
row even if the LP is in diving mode as long as the row does not
belong to the LP.
This is, e.g., useful for constraint handler or separator that
generate cuts based on a solution found by a diving heuristic.