add possibility to force restart if absolute gap is 1
If parameter forcerestart is true a restart is forced if the absolute
gap is 1 and the only variables that have nonzero (integral) objective
coefficients are indicator variables. This hopefully speeds up the
solution, since only infeasibiblity of the remaining system has to
be proved. Moreover, we switch to inference branching. The restart
is called from a new event handler.

The old event handler was renamed to make the name unique and the
implementation was moved to the top of the file.