use implications in reformulation of binary products
When reformulating a product y * (sum_i a_ix_i) for a binary variable
y, check whether there are implications from y=0 or y=1 when computing
bounds on the linear term (for the big-M).
Unfortunately, in my testset, this only applies to one instance,
icap6000, and there it does not make a real difference.

Similar, remove bilinear term x*y if x=1 implies y=0, or replace it by y
if x=0 implies y=0. But since this applies only to icap6000 and one
would need to add extra constraints to enforce this implication, leave
this off for now.