make storing of intermediate solutions a parameter
With Ipopt >= 3.10, the headers required to access solution values in
the intermediate callback are installed by Ipopt. So activate the
corresponding code and execute it whenever the user sets an appropriate
parameter. The parameter need to be set via an Ipopt options file, which
name can be set via the nlpi optfile parameter.

Unfortunately, the now included headers generate compiler warnings with
SCIP's choice of warnings flags for gcc.