added checking of LP solution status when trying root LP solution
This fixes bug 145 in bugzilla, where the time limit in the LP solver
is hit, but SCIP has still 8 seconds left and checks the LP solution.
This solution, however, was not updated since the current LP solution
was not primal feasible, so the old solution is checked and accepted,
but has infinite objective value.