new statistics for LPs without iterations
Introduced new counter and time measurements for accouting
of LPs that used no iterations (e.g. because "just" a refactorization)
was necessary. The statistics at the end now show the "total time"
used, the "total number of calls" of the primal/dual simplex or barrier.
Additionally at the end we have the time used for LPs without
iterations and the time used.