fixed assert in heur_rounding
For the intel compiler at some point 
minobj < SCIPgetCutoffbound(scip) is no longer true, although
it is previously tested in the variable selection. This has already
been noted earlier. It seems that this is due to some numerical
rounding problem at the last digit. The fix of the assert follows
the earlier suggestion and has no influence on the heuristic.