replace SCIPisFeasXX() by SCIPisXX() for SCIPvarGet[U|L]bAtIndex()
SCIPvarGet[U|L]bAtIndx() functions are usually used in reverse
propagation. The variable bounds at previous nodes seem to be
changed according to SCIPisXX() tests, i.e., even if the bound used to
satisfy SCIPisFeasXX() at the time of inference of the bound change
they are only guaranteed to satisfy it with SCIPisXX() later on.