enhanced syntax errors in reader_lp
The syntax error messages now end with the string passed to
the function syntaxError (the token has been moved to the front).
Consequently all strings now end with '.'.
When parsing indicator constraints now "<->" is detected and
a corresponding error message is displayed (these constraints
are currently not supported by SCIP).