use single function for strong branching calls in lp.c
There used to be two functions for strong branching. One for
integral and one for fractional branching. The code was the same
except for the corresponding call to the lpi. Now there is only
one function that takes a Bool to decide which function to call.
No functionality changes should have happend.