fixed bug 331, sepa_mcf, dynamic columns
For the current implementation of the MCF separator
variables and their LP positions should not change from one
call of separate to the next. In particular there should
not be new variables or deleted variables. We now test whether
nvars==ncols. If not the separator exits. Of course this is
not a proper fix.