added handling of integral variable to propagation of xor
cons_xor used to delete the integral variable with SCIPdelVar().
However, this leads to several issues, e.g., the variable might be
present in basis information etc.

The next idea was to fix the variable to the value given by the
other variables. However, this is sometimes not possible,
because this might be infeasible. Thus, one needs to take the
integral variable into account when propagating.

The propagation now takes the integral variable into account.
If all variables are fixed, it fixes the integral variable. If all but
one variable are fixed, this variable can be fixed as well due to
parity reasons. Then the integral variable can be fixed as well.
Furthermore, depending on the lower and upper bounds of the
integral variable one can fix variable to 0 and 1 respectively.

Furthermore, fixed two asserts in presolving (previous presolving
steps might have emptied the constraint and might have fixed
a variable chosen for fixation earlier).