change computation of gradient of sqrt
Previously, computing the square root at a point x
was done via sqrt(x + eps) - sqrt(eps) where eps = 1e-20
However this is not the correct behavior when computing gradient cuts:

Consider the following constraint
f(x,y) := -x - sqrt( y^2 ) <= -1
with bounds on y, 0 <= y

This is a convex constraint and the region represented by it is
{x,y | x+y>=1, y >= 0}. It looks like: \_
If we want to cut off (0,0), we build the gradient cut:
f(0) + gradf(0) ( (x,y) - (0,0) ) <= -1
Where gradf is (-1, 0)
because the derivative of -sqrt(y^2 + eps) equals -2y/(2 sqrt(y^2 + eps))
which evaluates to 0.
Then the cut is
-1 x <= -1
but this is invalid cut. It looks like: \_|->
and cuts off perfectly feasible points, like (0,1)

The problem is that while the original function
is convex, after perturbing the sqrt, it might not be convex anymore.
Exmaple: -sqrt(x^2) is convex but -sqrt(x^2 + eps) is not
(second derviate always negative)