fixed handling of Mosek
Mosek only supports shared libraries - this is now corrected.

Moreover, libmosek64 sets the soname to "libmosek64". Thus,
when running the binary, the dynamic loader searches for this name in
the specified path (<SCIPDIR>/lib/shared), but only finds the link to
Mosek not libmosek64. We fix this problem by adding an "rpath" option
that adds a link to the original location of the Mosek library.