remove superfluous chgupperbound variable
This variable was once needed, when a primal solution was generated
within extendToCover(). It could then happen that the objective improved,
which would again help to generate better cuts.
Since a while solutions are only passed to the corresponding heuristics.
Thus, this variable is a left over that it not needed anymore.